Moved date manipulation processing under utilities.
