+++ common/src/x509.ts
@@ -1,4 +1,5 @@
 import * as jsrsasign from "jsrsasign";
+import {addSeconds, formatDateTimeForDisplay, getCurrentUTCDate} from "./utils/datetime";
 
 export const CSR_PEM_PREAMBLE = "-----BEGIN CERTIFICATE REQUEST-----";
 export const CSR_PEM_POSTAMBLE = "-----END CERTIFICATE REQUEST-----";
@@ -41,27 +42,6 @@ export const createEcdsaCsr = (
   });
   return trimmer(regularCsr);
 };
-
-function getCurrentUTCDate(): Date {
-  return new Date(); // UTCで現在の日時を取得
-}
-
-function addSeconds(date: Date, secondsToAdd: number): Date {
-  const newDate = new Date(date);
-  newDate.setUTCSeconds(date.getUTCSeconds() + secondsToAdd);
-  return newDate;
-}
-
-function formatDateTimeForDisplay(date: Date): string {
-  const year = date.getUTCFullYear().toString().slice(-2);
-  const month = ("0" + (date.getUTCMonth() + 1)).slice(-2);
-  const day = ("0" + date.getUTCDate()).slice(-2);
-  const hours = ("0" + date.getUTCHours()).slice(-2);
-  const minutes = ("0" + date.getUTCMinutes()).slice(-2);
-  const seconds = ("0" + date.getUTCSeconds()).slice(-2);
-
-  return year + month + day + hours + minutes + seconds + "Z";
-}
 
 const positiveSerialNumber = () => {
   let hexValue: string;
